fix reputation mis type

+++ src-tauri/src/wfm_client/types/auction_owner.rs
@@ -8,7 +8,7 @@ pub struct AuctionOwner {
     pub last_seen: String,
 
     #[serde(rename = "reputation")]
-    pub reputation: i64,
+    pub reputation: f64,
 
     #[serde(rename = "locale")]
     pub locale: String,
